[xxx] Extract provider filter into services
