+++ app/controllers/support/allocations_controller.rb
@@ -19,7 +19,7 @@ module Support
     end
 
     def filter
-      @filter ||= ProviderFilter.new(params: filter_params)
+      @filter ||= Support::Providers::Filter.new(params: filter_params)
     end
 
     def filters
